moved footer test files into their directory
